requirements.txt
Flask
numpy
opencv-python-headless
pytesseract
pillow
fpdf2
Werkzeug